wip1

+++ rocket_avionics/telemetry.cpp
@@ -111,7 +111,8 @@ static uint16_t buildStateFlags() {
   if (flightIsArmReady()) flags |= (1 << 11); // [11] arm ready
   // [15:12] slot_seq_idx: slot_index % SLOT_SEQUENCE_LEN, wire-packing for passive sync.
   // SLOT_SEQUENCE_LEN must be ≤16 for this 4-bit field.
-  flags |= ((uint16_t)(radioGetSlotIndex() % SLOT_SEQUENCE_LEN) & 0x0F) << 12;
+  // Defensive: cast to uint64 before mod so negative slot indices (pre-anchor) wrap predictably.
+  flags |= ((uint16_t)(((uint64_t)radioGetSlotIndex()) % SLOT_SEQUENCE_LEN) & 0x0F) << 12;
   return flags;
 }
 
@@ -244,7 +245,7 @@ static void buildPage0C(uint8_t* buf, size_t* pos) {
   // Sync status and current slot index
   uint8_t syncFlags = radioInSync() ? 0x01 : 0x00;
   if (radioInSync()) {
-    uint8_t seqIdx = (uint8_t)(radioGetSlotIndex() % SLOT_SEQUENCE_LEN);
+    uint8_t seqIdx = (uint8_t)(((uint64_t)radioGetSlotIndex()) % SLOT_SEQUENCE_LEN);
     syncFlags |= (seqIdx & 0x7F) << 1;
   }
   writeU8(buf, pos, syncFlags);
